Prevent next_note to_fetch handler from running reactively
From [[statebus reactive side-effect]]:

Right now, I have  on the server functioning basically as an API call -- it doesn't actually store  state, instead running a function and returning the result reactively.

The problem is, it's subscribed to , meaning when I make a new note, it runs immediately and interacts with that new note in a way that is unnecessary and generates some bad UX.

This commit fixes this problem.