Move spaced repetition code to client-side
See [[statebus reactive side-effect]] for further notes. Excerpt:

I got it working with XHR, and it feels like the networking here is adding complexity pretty unnecessarily. Part of me feels like I'm violating some standard by moving this code to the client, like this isn't the normal way to do something like this, but... it would be easier, and I really don't see any reasons it would be a problem on medium horizon. It is local-first and relinquishes some ability to scrub bad inputs, but I don't see a need for that in the near term.

I'm switching over to local, and one downside is it's making a network request for every note in the `next_note` loop, which is a little slow. Would be better if the `next_rep`s were under the `all_notes` key so only one note needed to actually be fetched. I feel like there's probably another good solution with SB7 link keys, but that's for later. For now, I'm going to ignore this; if speed becomes a concern before SB7 is available (e.g. for putting the note server online) I'm going to see about moving `next_rep` into `all_notes`.

It does work locally, and I like this better for now.