Pick a random note from those scheduled same day
This is another stab at usability improvements around
[[spaced repetition on large batch annoying first passes in order]].
The idea is to break the monotony of going through the full corpus
multiple times by randomizing the order within a particular day.
This respects the recommendations of the SuperMemo2 algorithm, which
only operates at the resolution of days, while allowing the order to
be randomized in this particular scenario (many many new notes at
once).